VSL-98: Moved success modal from hook
