+++ packages/client/src/hooks/index.js
@@ -8,4 +8,3 @@ export { default as useFlowFees } from "./useFlowFees";
 export { default as useErrorMessage } from "./useErrorMessage";
 export { default as useAccount } from "./useAccount";
 export { default as useContacts } from "./useContacts";
-export { default as useSuccessMessage } from "./useSuccessMessage";
