+++ packages/client/src/pages/Safe.js
@@ -13,7 +13,8 @@ import {
 } from "../components";
 import { ArrowDown, ArrowUp } from "../components/Svg";
 import { Web3Consumer, useModalContext } from "../contexts";
-import { useClipboard, useErrorMessage, useSuccessMessage } from "../hooks";
+import { useClipboard, useErrorMessage } from "../hooks";
+import { TransactionSuccessModal } from "modals";
 
 const ReceiveTokens = ({ name, address }) => {
   const modalContext = useModalContext();
@@ -56,11 +57,10 @@ const ReceiveTokens = ({ name, address }) => {
 function Safe({ web3 }) {
   const params = useParams();
   const { address, tab } = params;
-  const modalContext = useModalContext();
+  const { openModal, closeModal } = useModalContext();
   const clipboard = useClipboard();
 
   const { showErrorModal } = useErrorMessage();
-  const { showTransactionSuccessModal } = useSuccessMessage();
 
   const safeData = web3?.treasuries?.[address];
   const actions = web3?.actions?.[address];
@@ -73,6 +73,22 @@ function Safe({ web3 }) {
           Couldn't find this safe's data...
         </div>
       </section>
+    );
+  }
+
+  const showTransactionSuccessModal = (action, safeName, safeAddress) => {
+    const actionData = action.data.actionView;
+    openModal(
+      <TransactionSuccessModal
+        safeName={safeName}
+        safeAddress={safeAddress}
+        actionData={actionData}
+        txID={action.transactionId}
+        onClose={closeModal}
+      />,
+      {
+        headerTitle: "Success",
+      }
     );
   }
 
@@ -181,13 +197,13 @@ function Safe({ web3 }) {
   const BodyComponent = tabMap[currentTab];
 
   const onSend = () => {
-    modalContext.openModal(
+    openModal(
       <SendTokens name={safeData.name} address={address} />
     );
   };
 
   const onReceive = () => {
-    modalContext.openModal(
+    openModal(
       <ReceiveTokens name={safeData.name} address={address} />
     );
   };
